// src/types/mock-data.type.ts
export type MockData = {
    header: string[];
    description: string[];
    publication_date: string[];
    preview: string[];
    photoes: string[][];
    type: string[];
    extras: string[];
    author: string[];
    city: string[];
    firstname: string[];
    email: string[];
    user_type: string[];
    avatar: string[];
}